C89 fixes

+++ src/tests/parse/test_integration_c2openapi.h
@@ -613,7 +613,7 @@ TEST test_c2o_cli_doc_sec_unset(void) {
   size_t i;
   for (i = 0; i < sizeof(snippets) / sizeof(snippets[0]); ++i) {
     char *c_file = NULL;
-    asprintf(&c_file, "%s%cf%zu.c", src_dir, PATH_SEP_C, i);
+    asprintf(&c_file, "%s%cf%lu.c", src_dir, PATH_SEP_C, (unsigned long)i);
     write_to_file(c_file, snippets[i]);
     free(c_file);
   }
@@ -625,7 +625,7 @@ TEST test_c2o_cli_doc_sec_unset(void) {
 
   for (i = 0; i < sizeof(snippets) / sizeof(snippets[0]); ++i) {
     char *c_file = NULL;
-    asprintf(&c_file, "%s%cf%zu.c", src_dir, PATH_SEP_C, i);
+    asprintf(&c_file, "%s%cf%lu.c", src_dir, PATH_SEP_C, (unsigned long)i);
     remove(c_file);
     free(c_file);
   }
@@ -906,7 +906,7 @@ TEST test_c2o_cli_oauth_validation_errors(void) {
   size_t i;
   for (i = 0; i < sizeof(snippets) / sizeof(snippets[0]); ++i) {
     char *c_file = NULL;
-    asprintf(&c_file, "%s%cf%zu.c", src_dir, PATH_SEP_C, i);
+    asprintf(&c_file, "%s%cf%lu.c", src_dir, PATH_SEP_C, (unsigned long)i);
     write_to_file(c_file, snippets[i]);
     free(c_file);
   }
@@ -918,7 +918,7 @@ TEST test_c2o_cli_oauth_validation_errors(void) {
 
   for (i = 0; i < sizeof(snippets) / sizeof(snippets[0]); ++i) {
     char *c_file = NULL;
-    asprintf(&c_file, "%s%cf%zu.c", src_dir, PATH_SEP_C, i);
+    asprintf(&c_file, "%s%cf%lu.c", src_dir, PATH_SEP_C, (unsigned long)i);
     remove(c_file);
     free(c_file);
   }
@@ -970,7 +970,7 @@ TEST test_c2o_cli_merge_oauth_flow_collisions(void) {
   size_t i;
   for (i = 0; i < sizeof(snippets) / sizeof(snippets[0]); ++i) {
     char *c_file = NULL;
-    asprintf(&c_file, "%s%cf%zu.c", src_dir, PATH_SEP_C, i);
+    asprintf(&c_file, "%s%cf%lu.c", src_dir, PATH_SEP_C, (unsigned long)i);
     write_to_file(c_file, snippets[i]);
     free(c_file);
   }
@@ -982,7 +982,7 @@ TEST test_c2o_cli_merge_oauth_flow_collisions(void) {
 
   for (i = 0; i < sizeof(snippets) / sizeof(snippets[0]); ++i) {
     char *c_file = NULL;
-    asprintf(&c_file, "%s%cf%zu.c", src_dir, PATH_SEP_C, i);
+    asprintf(&c_file, "%s%cf%lu.c", src_dir, PATH_SEP_C, (unsigned long)i);
     remove(c_file);
     free(c_file);
   }
